+++ examples/vue3/views/Home.vue
@@ -1,40 +1,38 @@
 <template>
-  <div>
-    <div class="filters">
-      <input type="checkbox" id="male" value="male" v-model="conditions.gender">
-      <label for="male">Male</label>
-      <input type="checkbox" id="female" value="female" v-model="conditions.gender">
-      <label for="female">Female</label>
-      <input class="date-picker" type="date" v-model="conditions.date">
-      <button class="btn" @click="refresh">Refresh</button>
-    </div>
-
-    <div class="container" v-if="!loading">
-      <div class="card" v-for="item in data.results" :key="item.id.value">
-        <div>
-          <img :src="item.picture.thumbnail"/>
-        </div>
-        <h4>{{`${item.name.first} ${item.name.last}`}}</h4>
-        <div>{{item.email}}</div>
-        <div>{{item.phone}}</div>
-      </div>
-    </div>
-    <div class="loading" v-else>Loading...</div>
+  <div class="filters">
+    <input type="checkbox" id="male" value="male" v-model="conditions.gender" :disabled="loading">
+    <label for="male">Male</label>
+    <input type="checkbox" id="female" value="female" v-model="conditions.gender" :disabled="loading">
+    <label for="female">Female</label>
+    <input class="date-picker" type="date" v-model="conditions.date" :disabled="loading">
+    <button class="btn" @click="execute">Refetch</button>
   </div>
+
+  <div class="container" v-if="!loading">
+    <div class="card" v-for="item in data.results" :key="item.id.value">
+      <div>
+        <img :src="item.picture.thumbnail"/>
+      </div>
+      <h4>{{`${item.name.first} ${item.name.last}`}}</h4>
+      <div>{{item.email}}</div>
+      <div>{{item.phone}}</div>
+    </div>
+  </div>
+  <div class="loading" v-else>Loading...</div>
 </template>
 
 <script lang="ts">
-import { defineComponent} from 'vue'
+import { defineComponent } from 'vue'
 import { useConditionWatcher } from '../../../src/index'
 // import { useConditionWatcher } from 'vue-condition-watcher'
 import api from '../api'
 
 export default defineComponent({
   setup(){
-
-    const { conditions, loading, data, refresh } = useConditionWatcher(
+    const { conditions, loading, data, execute } = useConditionWatcher(
       {
         fetcher: api.users,
+        initialData: {},
         defaultParams: {
           results: 9,
         },
@@ -44,10 +42,6 @@ export default defineComponent({
           offset: 0,
           limit: 9
         },
-        beforeFetch(conditions){
-          console.log(conditions)
-          return conditions
-        }
       }, 
       { 
         sync: 'router', 
@@ -59,8 +53,8 @@ export default defineComponent({
     return {
       loading,
       data,
-      refresh,
-      conditions
+      conditions,
+      execute,
     }
   }
 })