+++ src/types.ts
@@ -1,5 +1,7 @@
 import { Ref, InjectionKey, UnwrapRef } from 'vue-demi'
 
+type Fn = () => void
+
 export type UnwrapNestedRefs<T> = T extends Ref ? T : UnwrapRef<T>
 
 export type Conditions<T> = {
@@ -10,12 +12,20 @@ export type ConditionsType = {
   [key: string]: any
 }
 
+export interface OnFetchErrorContext<T = any, E = any> {
+  error: E
+  data: T | null
+}
+
 export interface Config<O> {
   fetcher: (params: object) => Promise<any>
+  immediate?: boolean
+  initialData?: any
   conditions: O
   defaultParams?: object
-  beforeFetch?: (conditions: O & object) => object
-  afterFetch?: (data: any) => void
+  beforeFetch?: (conditions: O & ConditionsType, cancel: Fn) => ConditionsType
+  afterFetch?: (data: any) => any
+  onFetchError?: (ctx: OnFetchErrorContext) => Promise<Partial<OnFetchErrorContext>> | Partial<OnFetchErrorContext>
 }
 
 export interface QueryOptions<K> {
@@ -24,10 +34,10 @@ export interface QueryOptions<K> {
   ignore?: Array<K>
 }
 
-export interface Result<O> {
+export interface UseConditionWatcherReturn<O> {
   conditions: UnwrapNestedRefs<O>
   loading: Ref<boolean | false>
   data: Ref<any | null>
-  refresh: Ref<() => void>
   error: Ref<any | null>
+  execute: () => Promise<any>
 }