+++ src/utils.ts
@@ -134,3 +134,7 @@ export function deepClone(obj): any {
     ? Array.from(obj)
     : clone
 }
+
+export function containsProp(obj: object, ...props: string[]) {
+  return props.some((k) => k in obj)
+}