+++ test/use-condition-watcher.test.ts
@@ -9,6 +9,12 @@ const tick = async (times: number) => {
   }
 }
 
+function doAsync(c) {
+  setTimeout(() => {
+    c()
+  }, 0)
+}
+
 describe('useConditionWatcher', () => {
   const root = document.createElement('div')
   const consoleWarnSpy = jest.spyOn(console, 'warn')
@@ -25,7 +31,7 @@ describe('useConditionWatcher', () => {
     expect(root).not.toBeNull()
   })
 
-  it(`Check return value's type`, async () => {
+  it(`Check return value's type`, () => {
     const config = {
       fetcher: (params) => new Promise((resolve) => resolve(params)),
       conditions: {
@@ -33,15 +39,14 @@ describe('useConditionWatcher', () => {
         results: 9,
       },
     }
-    const { conditions, data, error, loading, refresh } = useConditionWatcher(config)
+    const { conditions, data, error, loading, execute } = useConditionWatcher(config)
 
     expect(isReactive(conditions)).toBeTruthy()
     expect(isRef(data)).toBeTruthy()
     expect(isRef(error)).toBeTruthy()
     expect(isRef(loading)).toBeTruthy()
-    expect(isRef(refresh)).toBeTruthy()
     expect(typeof loading.value === 'boolean').toBeTruthy()
-    expect(typeof refresh.value === 'function').toBeTruthy()
+    expect(typeof execute === 'function').toBeTruthy()
   })
 
   it(`Condition should be change`, () => {
@@ -67,12 +72,13 @@ describe('useConditionWatcher', () => {
     })
   })
 
-  it(`Should return data from a promise`, async () => {
+  it(`Should return data from a promise`, () => {
     const vm = createApp({
       template: `<div>{{data}}</div>`,
       setup() {
         const config = {
           fetcher: () => new Promise((resolve) => resolve('ConditionWatcher')),
+          immediate: true,
           conditions: {
             gender: ['male'],
             results: 9,
@@ -82,12 +88,14 @@ describe('useConditionWatcher', () => {
       },
     }).mount(root)
 
-    expect(vm.$el.textContent).toBe('')
-    await tick(1)
-    expect(vm.$el.textContent).toBe('ConditionWatcher')
+    doAsync(async () => {
+      expect(vm.$el.textContent).toBe('')
+      await tick(1)
+      expect(vm.$el.textContent).toBe('ConditionWatcher')
+    })
   })
 
-  it(`Loading state should return true until promise resolve`, async () => {
+  it(`Loading state should return true until promise resolve`, () => {
     const vm = createApp({
       template: `<div>loading:{{loading}}, result:{{data}}</div>`,
       setup() {
@@ -102,14 +110,16 @@ describe('useConditionWatcher', () => {
       },
     }).mount(root)
 
-    await tick(1)
-    expect(vm.$el.textContent).toBe('loading:true, result:')
-    timeout(200)
-    await tick(1)
-    expect(vm.$el.textContent).toBe('loading:false, result:ConditionWatcher')
+    doAsync(async () => {
+      await tick(1)
+      expect(vm.$el.textContent).toBe('loading:true, result:')
+      timeout(200)
+      await tick(1)
+      expect(vm.$el.textContent).toBe('loading:false, result:ConditionWatcher')
+    })
   })
 
-  it(`Fetcher's params should same by condition and defaultParams`, async () => {
+  it(`Fetcher's params should same by condition and defaultParams`, () => {
     const vm = createApp({
       template: `<div>{{data}}</div>`,
       setup() {
@@ -128,16 +138,18 @@ describe('useConditionWatcher', () => {
       },
     }).mount(root)
 
-    await tick(1)
-    expect(JSON.parse(vm.$el.textContent)).toMatchObject({
-      results: 9,
-      name: 'runkids',
-      limit: 10,
-      offset: 1,
+    doAsync(async () => {
+      await tick(1)
+      expect(JSON.parse(vm.$el.textContent)).toMatchObject({
+        results: 9,
+        name: 'runkids',
+        limit: 10,
+        offset: 1,
+      })
     })
   })
 
-  it(`Fetcher's params should same with beforeFetch return object`, async () => {
+  it(`Fetcher's params should same with beforeFetch return object`, () => {
     const vm = createApp({
       template: `<div>{{data}}</div>`,
       setup() {
@@ -157,14 +169,16 @@ describe('useConditionWatcher', () => {
       },
     }).mount(root)
 
-    await tick(1)
-    expect(JSON.parse(vm.$el.textContent)).toMatchObject({
-      results: 9,
-      date: '2020-5-22',
+    doAsync(async () => {
+      await tick(1)
+      expect(JSON.parse(vm.$el.textContent)).toMatchObject({
+        results: 9,
+        date: '2020-5-22',
+      })
     })
   })
 
-  it(`If conditions attribute's type is array, fetcher's param should be string`, async () => {
+  it(`If conditions attribute's type is array, fetcher's param should be string`, () => {
     const vm = createApp({
       template: `<div>{{data}}</div>`,
       setup() {
@@ -178,7 +192,9 @@ describe('useConditionWatcher', () => {
       },
     }).mount(root)
 
-    await tick(1)
-    expect(vm.$el.textContent).toBe('male,female')
+    doAsync(async () => {
+      await tick(1)
+      expect(vm.$el.textContent).toBe('male,female')
+    })
   })
 })